Added options to config to more accurately reconstitute a torch model. - Paths to the ion dictionary, token dictionary, and yaml files that   created the model.
Not watching the data folder anymore.

The code placed in the models directory for creating the torch model is
very ad hoc. Need to figure out the __init__.py files.

Fixed an important bug in the shap_calculator.py that truncates strings
to a single character when subbing in the bgd tokens.
- Changed the output tensor data type to U12 instead of the default U1.